Fix issue #920 - changed the netlink parsing to use neli API, this will make sure endianess is handled properly

+++ tap/src/brokers/dot11_broker.rs
@@ -201,7 +201,7 @@ impl Dot11Broker {
         let is_wep = flags.map(|flags| flags.wep);
         
         let channel = if let Some(frequency) = frequency {
-            let channel = match dot11_frequency_to_channel(frequency) {
+            let channel = match dot11_frequency_to_channel(frequency as u32) {
                 Ok(c) => c,
                 Err(e) => {
                     warn!("Could not parse channel number from frequency: {}. Present Flags: {:?}", e, present_flags);
